refactor(collection): split 5157-line monolith into focused modules under src/commands/collection/
- Extracts shared helpers to src/functions/: ComponentsV2Utils, ImportCandidateUtils (IGDB search, candidate scoring, deduplication), resolveMemberLabel in InteractionUtils
- Splits collection into 11 files: steam-import service/command, csv-import service/command, autocomplete utils, game-resolve utils, import-ui utils, overview service, list service, crud command, view command
- collection.command.ts reduced to a 6-line barrel re-exporting buildCollectionListResponseForTests and the four command classes
- All slash group registrations remain under @SlashGroup('collection'); no command names or custom IDs changed
- tsc --noEmit passes with zero errors
