+++ src/commands/collection/collection-autocomplete.utils.ts
@@ -1,0 +1,72 @@
+import { type AutocompleteInteraction } from "discord.js";
+import Game from "../../classes/Game.js";
+import UserGameCollection, {
+  type CollectionOwnershipType,
+} from "../../classes/UserGameCollection.js";
+import { sanitizeUserInput } from "../../functions/InteractionUtils.js";
+import { formatGameTitleWithYear } from "../../functions/GameTitleAutocompleteUtils.js";
+
+export const COLLECTION_ENTRY_VALUE_PREFIX = "collection";
+
+export function parseCollectionEntryAutocompleteValue(raw: string): number | null {
+  const value = raw.trim();
+  const match = /^collection:(\d+)$/i.exec(value);
+  if (!match) return null;
+  const entryId = Number(match[1]);
+  if (!Number.isInteger(entryId) || entryId <= 0) return null;
+  return entryId;
+}
+
+export function buildCollectionEntryAutocompleteValue(entryId: number): string {
+  return `${COLLECTION_ENTRY_VALUE_PREFIX}:${entryId}`;
+}
+
+export function formatCollectionEntryAutocompleteName(entry: {
+  title: string;
+  platformName: string | null;
+  ownershipType: CollectionOwnershipType;
+}): string {
+  const platform = entry.platformName ?? "Unknown platform";
+  return `${entry.title} | ${platform} | ${entry.ownershipType}`.slice(0, 100);
+}
+
+export async function autocompleteCollectionGameTitle(
+  interaction: AutocompleteInteraction,
+): Promise<void> {
+  const focused = interaction.options.getFocused(true);
+  const rawQuery = focused?.value ? String(focused.value) : "";
+  const query = sanitizeUserInput(rawQuery, { preserveNewlines: false }).trim();
+  if (!query) {
+    await interaction.respond([]);
+    return;
+  }
+
+  const results = await Game.searchGamesAutocomplete(query);
+  await interaction.respond(
+    results.slice(0, 25).map((game) => ({
+      name: formatGameTitleWithYear(game).slice(0, 100),
+      value: String(game.id),
+    })),
+  );
+}
+
+export async function autocompleteCollectionEntry(
+  interaction: AutocompleteInteraction,
+): Promise<void> {
+  const focused = interaction.options.getFocused(true);
+  const rawQuery = focused?.value ? String(focused.value) : "";
+  const query = sanitizeUserInput(rawQuery, { preserveNewlines: false });
+
+  const results = await UserGameCollection.autocompleteEntries(
+    interaction.user.id,
+    query,
+    25,
+  );
+
+  await interaction.respond(
+    results.map((entry) => ({
+      name: formatCollectionEntryAutocompleteName(entry),
+      value: buildCollectionEntryAutocompleteValue(entry.entryId),
+    })),
+  );
+}
