+++ src/functions/ComponentsV2Utils.ts
@@ -1,0 +1,10 @@
+import { MessageFlags } from "discord.js";
+import { COMPONENTS_V2_FLAG } from "../config/flags.js";
+
+export function buildComponentsV2Flags(isEphemeral: boolean): number {
+  return (isEphemeral ? MessageFlags.Ephemeral : 0) | COMPONENTS_V2_FLAG;
+}
+
+export function buildComponentsV2EditFlags(): number {
+  return COMPONENTS_V2_FLAG;
+}
